+++ Exilence/Startup.cs
@@ -1,16 +1,13 @@
-﻿using Exilence.Contexts;
-using Exilence.Hubs;
+﻿using Exilence.Hubs;
 using Exilence.Interfaces;
 using Exilence.Repositories;
 using Exilence.Services;
 using Hangfire;
 using Hangfire.MemoryStorage;
-using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -30,12 +27,9 @@ namespace Exilence
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddApplicationInsightsTelemetry(Configuration);
-
             services.AddHangfire(c => c.UseMemoryStorage());
 
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-            //.AddJsonOptions(opts => { opts.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver(); });
 
             //Add distributed cache service backed by Redis cache
             services.AddDistributedRedisCache(options =>
@@ -56,10 +50,15 @@ namespace Exilence
                     });
             });
 
-            services.AddSignalR().AddStackExchangeRedis(Configuration.GetConnectionString("Redis"), options =>
-            {
-                options.Configuration.ChannelPrefix = "ExilenceSignalR";
-            });
+            services
+                .AddSignalR(hubOptions => {
+                    hubOptions.HandshakeTimeout = TimeSpan.FromSeconds(40);
+                })
+                .AddStackExchangeRedis(Configuration.GetConnectionString("Redis"), options =>
+                {
+                    options.Configuration.ChannelPrefix = "ExilenceSignalR";
+                    options.Configuration.ConnectTimeout = 10000;
+                });
 
             services.AddScoped<ILadderService, LadderService>();
             services.AddHttpClient<IExternalService, ExternalService>();
@@ -77,8 +76,6 @@ namespace Exilence
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
-                var telemetryConfig = app.ApplicationServices.GetService<TelemetryConfiguration>();
-                telemetryConfig.DisableTelemetry = true;
             }
 
             app.UseMvc();
@@ -95,13 +92,14 @@ namespace Exilence
             app.UseHangfireServer(hangfireOpts);
             app.UseHangfireDashboard();
 
-            BackgroundJob.Schedule<ILadderService>(ls => ls.UpdateLadders(), TimeSpan.FromMilliseconds(5000));
+            if (Configuration.GetValue<bool>("Ladder:PollingEnabled")) { 
+                BackgroundJob.Schedule<ILadderService>(ls => ls.UpdateLadders(), TimeSpan.FromMilliseconds(5000));
+            }
 
             app.UseSignalR(routes =>
             {
                 routes.MapHub<PartyHub>("/hubs/party", options =>
                 {
-                    // 30Kb message buffer
                     options.ApplicationMaxBufferSize = 50 * 1024 * 1024;
                     options.TransportMaxBufferSize = 50 * 1024 * 1024;
                 });