+++ ExilenceClient/main.ts
@@ -1,7 +1,6 @@
 import { app, BrowserWindow, dialog, globalShortcut } from 'electron';
 import * as path from 'path';
 import * as url from 'url';
-import * as fs from 'fs';
 
 
 export interface ExileWindowEvent {
@@ -26,37 +25,19 @@ const windows: BrowserWindow[] = [];
 const args = process.argv.slice(1);
 const serve = args.some(val => val === '--serve');
 
-ipcMain.on('keybinds-update', function (event, binds) {
-  globalShortcut.unregisterAll();
-  globalShortcut.register('Command+Shift+I', () => {
-    windows[ExileWindowEnum.Main].openDevTools();
-  });
-  binds.forEach(bind => {
-    if (bind.enabled) {
-      globalShortcut.register(bind.keys, () => {
-        windows[ExileWindowEnum.Main].webContents.send('keybind', bind);
-      });
-    }
-  });
-});
-
-ipcMain.on('keybinds-unregister', function (event) {
-  globalShortcut.unregisterAll();
-});
-
 ipcMain.on('popout-window-update', (event, window: ExileWindowEvent) => {
   if (windows[window.event] && !windows[window.event].isDestroyed()) {
     windows[window.event].webContents.send('popout-window-update', window);
   }
 
-  if (window.event === ExileWindowEnum.Networth) {
-    const stream = fs.createWriteStream('networth.txt');
-    stream.once('open', function () {
-      stream.write(`Networth: ${Math.round(window.data.networth * 10) / 10}\r\n`);
-      stream.write(`Gain: ${Math.round(window.data.gain * 10) / 10}\r\n`);
-      stream.end();
-    });
-  }
+  // if (window.event === ExileWindowEnum.Networth) {
+  //   const stream = fs.createWriteStream('networth.txt');
+  //   stream.once('open', function () {
+  //     stream.write(`Networth: ${Math.round(window.data.networth * 10) / 10}\r\n`);
+  //     stream.write(`Gain: ${Math.round(window.data.gain * 10) / 10}\r\n`);
+  //     stream.end();
+  //   });
+  // }
 
 });