+++ ExilenceClient/src/app/authorize/authorize.module.ts
@@ -7,19 +7,18 @@ import { RouterModule } from '@angular/router';
 
 import { AnalyticsService } from '../shared/providers/analytics.service';
 import { IncomeService } from '../shared/providers/income.service';
-import { KeybindService } from '../shared/providers/keybind.service';
 import { MessageValueService } from '../shared/providers/message-value.service';
 import { PartyService } from '../shared/providers/party.service';
+import { PricingService } from '../shared/providers/pricing.service';
 import { SharedModule } from '../shared/shared.module';
 import { AuthorizeComponent } from './authorize.component';
+import { ServerMessageDialogComponent } from './components/server-message-dialog/server-message-dialog.component';
+import { ServerMessageDialogModule } from './components/server-message-dialog/server-message-dialog.module';
 import { DashboardModule } from './dashboard/dashboard.module';
 import { FaqModule } from './faq/faq.module';
 import { InspectPlayersModule } from './inspect-players/inspect-players.module';
 import { PartyModule } from './party/party.module';
 import { SettingsModule } from './settings/settings.module';
-import { PricingService } from '../shared/providers/pricing.service';
-import { ServerMessageDialogComponent } from './components/server-message-dialog/server-message-dialog.component';
-import { ServerMessageDialogModule } from './components/server-message-dialog/server-message-dialog.module';
 
 @NgModule({
   imports: [
@@ -43,7 +42,6 @@ import { ServerMessageDialogModule } from './components/server-message-dialog/se
   ],
   declarations: [AuthorizeComponent],
   providers: [
-    KeybindService,
     MessageValueService,
     PartyService,
     PricingService,