+++ ExilenceClient/src/app/authorize/components/stashtab-list/stashtab-list.component.ts
@@ -1,6 +1,6 @@
 import { SelectionModel } from '@angular/cdk/collections';
 import { Component, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { MatSort, MatTableDataSource } from '@angular/material';
+import { MatDialog, MatSort, MatTableDataSource } from '@angular/material';
 import { Subscription } from 'rxjs';
 
 import { Stash, Tab } from '../../../shared/interfaces/stash.interface';
@@ -8,6 +8,7 @@ import { AlertService } from '../../../shared/providers/alert.service';
 import { ExternalService } from '../../../shared/providers/external.service';
 import { PartyService } from '../../../shared/providers/party.service';
 import { SettingsService } from '../../../shared/providers/settings.service';
+import { MaptabInfoDialogComponent } from './maptab-info-dialog/maptab-info-dialog.component';
 
 @Component({
   selector: 'app-stashtab-list',
@@ -30,7 +31,8 @@ export class StashtabListComponent implements OnInit, OnDestroy {
     private settingsService: SettingsService,
     private externalService: ExternalService,
     private partyService: PartyService,
-    private alertService: AlertService
+    private alertService: AlertService,
+    private maptabDialog: MatDialog
   ) { }
 
   ngOnInit() {
@@ -41,7 +43,6 @@ export class StashtabListComponent implements OnInit, OnDestroy {
   }
 
   init() {
-    const sessionId = this.settingsService.get('account.sessionId');
     const accountName = this.settingsService.get('account.accountName');
     const league = this.partyService.currentPlayer.character.league;
     let selectedStashTabs: any[] = this.settingsService.get('selectedStashTabs');
@@ -50,7 +51,7 @@ export class StashtabListComponent implements OnInit, OnDestroy {
       selectedStashTabs = [];
     }
 
-    this.stashTabSub = this.externalService.getStashTabs(sessionId, accountName, league)
+    this.stashTabSub = this.externalService.getStashTabs(accountName, league)
       .subscribe((res: Stash) => {
         if (res !== null) {
           this.dataSource = res.tabs.map((tab: Tab) => {
@@ -95,8 +96,24 @@ export class StashtabListComponent implements OnInit, OnDestroy {
     }
   }
 
+  openMaptabDialog(): void {
+    const dialogRef = this.maptabDialog.open(MaptabInfoDialogComponent, {
+      width: '1100px'
+    });
+    dialogRef.afterClosed().subscribe(result => {
+    });
+  }
+
   showAlert() {
     this.alertService.showAlert({ message: 'You can select at most 20 stash tabs', action: 'OK' });
+  }
+
+  shouldToggle(selection, row) {
+    const mapTabSelected = selection.selected.find(x => x.position === row.position) !== undefined;
+    if (row.isMapTab && !mapTabSelected) {
+      this.openMaptabDialog();
+    }
+    this.toggle(selection, row);
   }
 
   toggle(selection, row) {