+++ ExilenceClient/src/app/shared/providers/external.service.ts
@@ -1,12 +1,16 @@
+import 'rxjs/add/operator/delay';
 import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/retryWhen';
+import 'rxjs/add/operator/take';
 import 'rxjs/add/operator/toArray';
+import 'rxjs/operators/mergeMap';
 
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
+import RateLimiter from 'rxjs-ratelimiter';
 import { BehaviorSubject } from 'rxjs/internal/BehaviorSubject';
 import { Observable } from 'rxjs/Observable';
-
 import { AccountInfo } from '../interfaces/account-info.interface';
 import { EquipmentResponse } from '../interfaces/equipment-response.interface';
 import { Item } from '../interfaces/item.interface';
@@ -18,7 +22,7 @@ import { Stash } from '../interfaces/stash.interface';
 import { AnalyticsService } from './analytics.service';
 import { ElectronService } from './electron.service';
 import { LogService } from './log.service';
-
+import { RequestHelper } from '../helpers/request.helper';
 
 @Injectable()
 export class ExternalService {
@@ -26,6 +30,12 @@ export class ExternalService {
 
   public leagues: BehaviorSubject<League[]> = new BehaviorSubject<League[]>([]);
   public tradeLeagueChanged = false;
+
+  private TradeSearchRequestLimit = new RateLimiter(1, 1200);
+  private TradeFetchRequestLimit = new RateLimiter(1, 600);
+
+  // combined ratelimiter for stash- and character-requests (singleton)
+  private RequestRateLimit = new RateLimiter(7, 10000);
 
   constructor(
     private http: HttpClient,
@@ -40,23 +50,19 @@ export class ExternalService {
   }
 
   getCharacter(data: AccountInfo): Observable<any> {
-    this.setCookie(data.sessionId);
-
     const parameters = `?accountName=${data.accountName}&character=${data.characterName}`;
 
-    return this.http.get('https://www.pathofexile.com/character-window/get-items' + parameters, { withCredentials: true }).catch(e => {
-      if (e.status !== 403 && e.status !== 404) {
-        this.logService.log('Could not character items, disconnecting!', null, true);
-        this.router.navigate(['/disconnected', true]);
-      }
-      return Observable.of(null);
-    });
+    return this.RequestRateLimit.limit
+      (this.http.get('https://www.pathofexile.com/character-window/get-items' + parameters, { withCredentials: true }).catch(e => {
+        if (e.status !== 403 && e.status !== 404) {
+          this.logService.log('Could not character items, disconnecting!', null, true);
+          this.router.navigate(['/disconnected', true]);
+        }
+        return Observable.of(null);
+      }));
   }
 
   getCharacterList(account: string, sessionId?: string) {
-    if (sessionId !== undefined) {
-      this.setCookie(sessionId);
-    }
 
     const parameters = `?accountName=${account}`;
     return this.http.get('https://www.pathofexile.com/character-window/get-characters' + parameters)
@@ -81,9 +87,7 @@ export class ExternalService {
       });
   }
 
-  getStashTabs(sessionId: string, account: string, league: string) {
-    this.setCookie(sessionId);
-
+  getStashTabs(account: string, league: string) {
     const parameters = `?league=${league}&accountName=${account}&tabs=1`;
     return this.http.get<Stash>('https://www.pathofexile.com/character-window/get-stash-items' + parameters)
       .catch(e => {
@@ -95,11 +99,14 @@ export class ExternalService {
       });
   }
 
-  getStashTab(sessionId: string, account: string, league: string, index: number): Observable<Stash> {
+  getStashTab(account: string, league: string, index: number): Observable<Stash> {
     this.analyticsService.sendEvent('income', `GET Stashtab`);
-    this.setCookie(sessionId);
     const parameters = `?league=${league}&accountName=${account}&tabIndex=${index}&tabs=1`;
-    return this.http.get<Stash>('https://www.pathofexile.com/character-window/get-stash-items' + parameters)
+    return this.RequestRateLimit.limit(
+      this.http.get<Stash>('https://www.pathofexile.com/character-window/get-stash-items' + parameters))
+      .retryWhen((error) => {
+        return error.delay(1000).take(3);
+      })
       .catch(e => {
         if (e.status !== 403 && e.status !== 404) {
           this.logService.log('Could not fetch stashtabs, disconnecting!', null, true);
@@ -110,7 +117,9 @@ export class ExternalService {
   }
 
   validateSessionId(sessionId: string, account: string, league: string, index: number) {
+    this.removeCookie();
     this.setCookie(sessionId);
+
     const parameters = `?league=${league}&accountName=${account}&tabIndex=${index}&tabs=1`;
     return this.http.get<Stash>('https://www.pathofexile.com/character-window/get-stash-items' + parameters)
       .catch(e => {
@@ -127,57 +136,74 @@ export class ExternalService {
 
   getAccountForCharacter(character: string) {
     const parameters = `?character=${encodeURIComponent(character)}`;
-    return this.http.get('https://www.pathofexile.com/character-window/get-account-name-by-character' + parameters);
+    return this.RequestRateLimit.limit
+      (this.http.get('https://www.pathofexile.com/character-window/get-account-name-by-character' + parameters));
   }
 
-  getPublicMapsFromTradeIds(subLines: any[], query: string) {
+  FetchPublicMaps(subLines: any[], query: string) {
     return Observable.from(subLines)
       .concatMap((lines: any) => {
         const url = `https://www.pathofexile.com/api/trade/fetch/${lines.join(',')}?query=${query}`;
-        return this.http.get(url)
-          .delay(1000);
+        return this.TradeFetchRequestLimit.limit(
+          this.http.get(url).retryWhen((err) => {
+            return err.flatMap((error: any) => {
+              if (error.status === 429) {
+                return Observable.of(error.status).delay(1000 * 10); // 10 second retry
+              }
+              return Observable.of(null);
+            }).take(2);
+          })
+        );
       }).toArray();
   }
 
-  getPublicMapTradeGuids(account: string, league: string) {
+  SearchPublicMaps(account: string, league: string) {
 
-    const requestUrl = 'https://www.pathofexile.com/api/trade/search/' + league;
-    const requestJson = {
-      'query': {
-        'status': {
-          'option': 'any'
-        },
-        'stats': [{
-          'type': 'and',
-          'filters': [],
-          'disabled': true
-        }],
-        'filters': {
-          'trade_filters': {
-            'disabled': false,
+    const tierObservable = [17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1];
+
+    return Observable.from(tierObservable)
+      .concatMap((tier: any) => {
+
+        const requestUrl = 'https://www.pathofexile.com/api/trade/search/' + league;
+        const requestJson = {
+          'query': {
+            'status': {
+              'option': 'any'
+            },
             'filters': {
-              'account': {
-                'input': account
+              'trade_filters': {
+                'disabled': false,
+                'filters': {
+                  'account': {
+                    'input': account
+                  }
+                }
+              },
+              'map_filters': {
+                'disabled': false,
+                'filters': {
+                  'map_tier': {
+                    'min': tier,
+                    'max': tier
+                  }
+                }
               }
             }
           },
-          'map_filters': {
-            'disabled': false,
-            'filters': {
-              'map_tier': {
-                'min': 0
-              }
-            }
+          'sort': {
+            'price': 'asc'
           }
-        }
-      },
-      'sort': {
-        'price': 'asc'
-      }
-    };
-
-    return this.http.post(requestUrl, requestJson);
-
+        };
+        return this.TradeSearchRequestLimit.limit(this.http.post(requestUrl, requestJson)
+          .retryWhen((err) => {
+            return err.flatMap((error: any) => {
+              if (error.status === 429) {
+                return Observable.of(error.status).delay(1000 * 60); // One minute retry
+              }
+              return Observable.of(null);
+            }).take(2);
+          }));
+      });
   }
 
   setCookie(sessionId: string) {
@@ -190,10 +216,20 @@ export class ExternalService {
       secure: false,
       httpOnly: false,
       expirationDate: undefined
-    };
+    } as Electron.Details;
 
     this.electronService.remote.session.defaultSession.cookies.set(cookie, (error) => {
-      if (error) { console.error(error); }
+      if (error) {
+        this.logService.log('Could not set cookie', error, true);
+      }
+    });
+  }
+
+  removeCookie() {
+    this.electronService.remote.session.defaultSession.cookies.remove('http://www.pathofexile.com', 'POESESSID', (error) => {
+      if (error) {
+        this.logService.log('Could not set cookie', error, true);
+      }
     });
   }